+++ carp_webservices/test/credentials.dart
@@ -1,0 +1,4 @@
+final String clientID = "carp";
+final String clientSecret = "carp";
+final String username = "study1own1er@cachet.dk";
+final String password = "researcher";